Dialogs: implementation prop renamed to contentItem; API cleanup
- contentItem is a common name for an item inside something such as a
window or a flickable.  We now consider it adequate API that you can
directly set this to some item which you want to fill the whole dialog
instead of adding an item as a child of the default property and being
subject to the dialog's margins and space reservation for buttons.
- contentItem MUST be an Item.  Allowing the old "implementation"
property to alternatively be a ready-made Window was just a
hypothetical use case, but if anyone made use of that, then it would
be a problem to have that kind of dialog on a GUI that cannot have
windows (such as Android or EGLFS). So there's no point in having
the possibility as long as we cannot emulate windows as items.
- standardButtonsRightModel and standardButtonsLeftModel are
implementation details so they should not be exposed as properties
which Creator can discover; users shouldn't depend on them.

Task-number: QTBUG-38056
Change-Id: Iefa3def314353495cfe8d1ef9f775a8e46d5bcf4
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>