Ensure that styleData.row is initialized in delegates
While we specifically documented that you should not depend on
the value being correct in onCompleted for 5.3, it is not
optimal that the initial value is consistently set to 0 and later
changed to it's appropriate value. This patch ensures that delegates
are not instantiated until after the model model and row values
have been fully initialized.

Task-number: QTBUG-38307
Change-Id: Ib3300539a0283c997ab8b808fb229d3f22ed6324
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>