Fix missing tabs in touch example
The only reason that the we calculated the tab count
explicitly before was that we were using an array.
Now that we are using a model there is no longer any
need to keep track of the count independently of the model.

This fix prevents the count from originally showing as 0 and
prevents an extra value change. In addition the count
is now correctly marked as readonly.

Task-number: QTBUG-38421
Change-Id: I53a5d0df3367c850a7c6846729660dd0f63238d5
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>
Reviewed-by: Jonathan Liu <net147@gmail.com>