Menu: Fix QPA related crash on exit
In QPA world, menus are always referenced by a menu item. (This
is the menu item either on the menu bar or in the parent menu.)
So, the same way that we set the menu item's menu in the menu's
constructor, we must clear it. Specially since qtbase commit
8605f44097a986e10c2ddaf4b4eae0b2331d98d1 which checks for the
menu reference in the menu item's destructor.

Change-Id: Id196fc7c2082dbff1d9006cbd82b7237532c4e3c
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>