FileDialog: update icon whenever the filename changes
view.model.get(styleData.row, "fileIsDir") is unreliable for some
reason, but each filename usually changes when we go into a different
directory.  Maybe it could still fail if you navigate from a directory
containing a directory, to another directory which has a file with
exactly the same name in the same position.  Since most files have
extensions, this is unlikely...

Task-number: QTBUG-38683
Change-Id: If080c4f1a7a01b20f6c468fcb06653e3c2314a1a
Reviewed-by: Liang Qi <liang.qi@digia.com>
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>