Don't allow navigation to previous/next month when dragging mouse.
When dragging the mouse over the calendar and a date from a previous
month is hit, it will change the visible month to that month.
This becomes worse when dragging the mouse between in the start/end of
the calendar, because there are usually several days shown from the
previous month, and the calendar quickly goes back several months.

[ChangeLog][Calendar] Calendar no longer selects dates
in the next/previous month when dragging the mouse.

Task-number: QTBUG-38848
Change-Id: I0b9bc727a05397a8e4f1280e16c91e39e160bf4a
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>