Add styleData.hasSelection for text selection handles
Text selection handles typically look different based on whether the
editor has selection or not (one handle for moving the cursor vs. two
handles for selecting text). SpinBox and ComboBox do not have suitable
properties for testing whether the control has selection or not, so
styleData.hasSelection is introduced to abstract and expose that
information.

Task-number: QTBUG-38934
Change-Id: I362b5790f5d487e5ba258e42d1b801a5cce76770
Reviewed-by: Liang Qi <liang.qi@digia.com>