ComboBox: respect style padding for the editor
Get rid of the hard coded 8px left margin and use whatever value
the style supplies (and is already used for the label). This allows
the Android style to provide a sensible value for High DPI screens.

The new value for the left margin is 6px in Base and Desktop styles.
It looks slightly better, but still not pixel perfect. Notice that
the text label of editable and non-editable ComboBoxes are not
perfectly aligned in native Windows and OS X ComboBoxes either.

Change-Id: I9a66c3a26fcd9ef1653b355e831277a7b6cbe72f
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>