Fix TextHandle min/max position handling
When not explicitly specified, the maximum is the length of the editor
content, not minus one (cursor can be moved "past" the last character).

Furthermore, due to binding evaluation orders, the minimum/maximum
properties might be temporarily pointing outside the range so we must
validate them to avoid clients (TextField/TextArea) having to do that.

Change-Id: I1faddd1e7ac4f12d7a802484ceb8eea500ad8bfc
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>