QQuickMenu: add enum MenuType that reflects QPlatformMenu::MenuType
On mobile platforms several different popup menu types exist.
E.g on iOS, you have a special popup just for selecting text.
A recent patch added to QtBase added a new enum to QPlatformMenu
that lets UI controls select/hint which one to use. This patch
adds the same enum to QQuickMenu so that we can control this
from Controls as well.

Change-Id: Ibf4bdc84577b5a6527021b1cb15578b56bb4a92e
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>