TableViewStyle: Add header delegate for Android style
The whole TableView concept doesn't map very nicely on Android
(and in mobile in general). However, for completeness, we provide
a fully implemented TableViewStyle.

For the header delegate, we re-used the design for section headers
which is close enough, specially since it's not uncommon to see
sticky section headers in mobile applications.

This change depends on another change in qtbase, which takes care
of extracting the 'listSeparatorTextViewStyle' element.

Change-Id: I9c50e4c7cabdc6b0ea864cef3da55ae3d7337d79
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>