FileDialog: ensure that arrow keys can always navigate the file list
By default the file list should be focused so that arrow keys can be
used to select a file, and so that the list has a focused appearance.
But even when the list does not have focus, the up/down arrow keys
can still navigate the list because they are not accepted by any
other controls in this dialog.

Change-Id: Ie73bb187672cce86da0e5089d16a558af92414e9
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>