Fix ApplicationWindow state / content size
QML Window is not a plain QQuickWindow, but QQuickWindowQmlImpl that
does delayed visibility handling and shows the window in appropriate
state. Thus, ApplicationWindow cannot inherit QQuickWindow or it would
lose all that magic.

QQuickApplicationWindow was added in order for ApplicationWindowStyle
to be able to know whether a custom window color had been set. This
change gets rid of the QQuickWindow subclass and simply compares the
window color to white, which is the default fill color of QQuickWindow.

Change-Id: Ibd95ef0717c2db51246c6a71585e9286e61cef0b
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@digia.com>
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>