Dialog: initialize enabled and clicked buttons
When the dialog becomes visible, no button has been clicked yet: we
are waiting for the user to do that.  Failure to initialize
m_clickedButton made the logic in accept() and reject() unreliable:
accept() does not imply OK unless we are sure the user did not click
any other button, and the same with reject() implying Cancel.

Change-Id: I340c18bd0fca6cc9a7274c1063882de425444326
Reviewed-by: Liang Qi <liang.qi@digia.com>