FileDialog: add sidebarVisible property and button
The sidebar takes up too much space on some devices, so now it is
possible for the application author to hide it by default.  The user
can also toggle it with a button similar to that used for the
Qt Creator sidebar, and this state is stored as a setting; however
if the application sets sidebarVisible: false, that overrides the setting.

[ChangeLog][QtQuick.Dialogs] FileDialog: added sidebarVisible property
and button

Task-number: QTBUG-39231
Change-Id: Ie8b379da02f4b12764732b2c01b79039722778ab
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>