FileDialog: tweak initial splitter positions
In Component.onCompleted, shortcutsView.model.count is undefined, so
the height of the shortcuts area was dictated by its minimum.  Further,
the ListView removes delegates before they are completely "below the
fold", so at a height of 2.5 rows, you see only two rows.  The
intention is to have the lowest item that you initially see
be visually cut off, as a hint that there are more items.
The sidebar was also too narrow by default on screens that have space
for it, so the default width is increased (but as before, the user's
preferred location for the splitter is saved to settings and restored
each time).

Change-Id: I74dc924254f0c87e6200c3fff6af900ff28b5b6a
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>