QtQuick.Dialogs: install translator if suitable translations are found
qt_XX.qm is only for Qt 4 compatibility and therefore does not link
to qtquickcontrols_XX.qm.  The fact that QQmlApplicationEngine
loads qt_XX.qm is wrong, insufficient to support dialog translations,
and also inefficient in that it loads some translations which we
don't need.  The loading of relevant translations should be modular
along the same lines that Qt is modular.  Each Qt module needs
to load its own translations.

In this case it's needed so that buttons on dialogs will be translated.
Dialogs depends on Controls, and the translation file is named as the
controls translation file, so any future Controls translations can go
here too.  It's more efficient to load fewer files than to load separate
translations for each QML module.

Change-Id: I5fe0c39643f066b15a605b53158e9185792fd33e
Reviewed-by: Jan Arve Sæther <jan-arve.saether@digia.com>