Fix asynchronous loading of Calendar.
For some reason, adding the days of the week to the header model on
component completion caused some (six) of the delegates to never be
loaded.

Instead, we assume that for most cases, the locale will be en_US,
and only change the model via setProperty if this is not the case.
This is enough to cause the delegates to be loaded properly.

Change-Id: Ic399be4b3c9e252fec4eafac84bd9d44c61c1150
Task-number: QTBUG-41481
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>