qquickmenu: add support for providing a target rect to __popup()
Using a target rect as menu location instead of a position has
been supported in QPlatformMenu for a while. The reason for
specifying a rect instead of a position is that then the OS
can decide if the popup should be placed above or below the
target rect so that it fits inside the screen.

A typical example is when showing an edit menu around a
text selection. If the selection (target rectangle) is
far up on the screen, the popup (with arrow) will be
placed below the selection instead of above, which is
the normal.

Change-Id: Ie6cd6a92f1d9ef480c1e455960021c7f18f86569
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>