iOS style: let TextAreaStyle inherit ScrollViewStyle
TextAreaStyle (iOS) inherits TextAreaStyle (base).
And TextAreaStyle (base) inherits ScrollViewStyle.
The problem is that since TextAreaStyle (base) is
in the base folder, it will also choose the ScrollViewStyle
from that folder as subclass. But what we want
is to use the ScrollViewStyle from the iOS folder.

It seems like the correct way is to not let TextAreaStyle (iOS)
inherit TextAreaStyle (base), but instead ScrollViewStyle (iOS)
directly. This means that we also need to redefine the same
properties in TextAreaStyle (iOS) as in TextAreaStyle (base)

Change-Id: If2ea16c5c334a0c5858cef1dc7886757fdbf2bd9
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>