Dialogs: added private maximumDimension property for initial size
The Screen.attached property does not know which screen will be used
until the dialog is shown, but we need to know how big the desktop is
before it is shown in order to constrain a reasonable dialog size,
before the Flow lays out the buttons.  Otherwise it will be too tall.

Task-number: QTBUG-41734
Change-Id: I44349fab4ad0254d2c6faf30c759fb9e52eebf03
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>