TextArea: ensure selection handles are visible when inside the flickable
The current implementation would hide the handles when parented
into the flickable and the content view was flicked up. The handles
would in those cases have an y pos greater than the viewport height. But
since contentY would be greater as well, the handles should still
be visible.

This patch will ensure that the visible logic takes the handle parent
into account when calculating if it overlaps the viewport.

Change-Id: I591a4eb8a71b89720c80b190113888b5601fd060
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>