Make the cursor and selection handle delegates private for now
The API is not ideal, we don't want to lock this thing down like as it
is now. All styles implemented so far created CursorHandleStyle.qml and
SelectionHandleStyle.qml so perhaps that's the better way to go.

Furthermore, there's never going to be multiple occurrences of the
handle pairs, so it should be one shared instance. Thus, the relevant
style delegates probably shouldn't be sprinkled around various control
styles...

So, for now, the most pragmatic solution is to make the handles and
cursors internal since we don't have time to revamp it before 5.4.0.

Change-Id: Ib9fcc4a431bb510c8c55a760159f4d6a631554b3
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@theqtcompany.com>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@digia.com>