EditMenu iOS: show edit menu upon pressnhold on empty input field
If you press and hold on an input field that doesn't have focus, we
select the word under the cursor, which will make the menu show.
But if the field is empty, no word can be selected, so therefore no
menu will show.
This patch will check for this special case, and show the menu if
you pressnhold on an empty input field.

Change-Id: I7d2c137765471cb1ff1dc4f7e34577e5e1317dc0
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>