Fix text selection on desktop
So far, text selection handles were only provided by mobile
specific styles. The new Flat style is an exception that
provides text selection handles for mobile, but is still
intended to be usable on desktop too.

The text selection handling code thus needs not only to check
whether the text selection handles exist, but also whether it's
a mobile platform. This fixes text selection on desktop.

Change-Id: Ibda2045ca42970e7db692d2e8370ac0dba21625a
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>