Add autotest for baselines
The heuristic of finding the visual baseline is based on analyzing the
rendered pixmap.
This is not pretty, but we really need to have a test for baselines
in order to not cause future regressions whenever things are changed in
the styles.

This is the reason why we choose to use a string of 'L's, since
the letter has a distinctive baseline.

If it turns out that the heuristic gives too many false
positives, this approach should be reconsidered.

Change-Id: Ic8de576244fa032617ff7546ddbe20dc6910c7bf
Reviewed-by: Caroline Chao <caroline.chao@theqtcompany.com>