Add support for Layout margins
Currently, the way to achieve this is to make a proxy item in a layout:
RowLayout {
    Item {
	Text {
	    anchors.fill: parent
	    anchors.margins: 9
	}
    }
}

using Layout margins you can now skip the proxy item and instead write:

RowLayout {
    Text {
        Layout.margins: 9
    }
}

Note that this is not the same as margins in the QLayout world, as this
only controls the margins between the layout container and its items.
This feature adds margins to each individual item inside the layout

The real logic is done by the container classes:
Layouts has the additional logic to adhere to them.

SplitView will be done in a separate commit.

Task-number: QTBUG-39724
Task-number: QTBUG-41559
Change-Id: I9c5c2a1e0ee43c1200b62009216c9f64d2875baf
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>