BasicTableView refactoring: Separate view from user interaction
One of the main tasks of the user interaction code is to manage
the current items selection. However, in the upcoming TreeView,
the selection will be completely different, but we still want to
keep 99% of the view code.

Also, in our TreeView implementation, the user's model is proxyfied
by a TreeModelAdaptor. But that's not the case in the TableView.
So, in order to keep using the same API, we introduce the __model
intermediate property.

Where it makes sense, TableView and TreeView share common API
through BasicTableView.

Change-Id: I0a922c1c1093919df7d880d72fe12ed0f372cc08
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>