Use all, not just available geometry in menu popup
A menu popup may be positioned partly or fully outside
of the available geometry. For example if a Windows
side bar (appbar) is opening such a popup, it will be
completely outside of the available geometry (rcWork from
MONITORINFOEX) since such sidebars are not part of that.

Instead, use the regular, full geometry.

Task-number: QTBUG-43693
Change-Id: Ibd22150cb7ec94a46df2facbfc1bf5f20bda51e5
Reviewed-by: Shawn Rutledge <shawn.rutledge@digia.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>