Dialogs: implement QML backup support for native file dialog
QtQuickAbstractDialog will fall back to use QML version of a dialog
if a dialog helper returns false on show. We will utilize this for
file dialogs by letting a platform be selective about when to show
a native dialog. A platform can use this to e.g show specific native
dialogs for certain dialog configurations, and otherwise return false
to signal that the caller should fall back to use the QML/widgets version.

This patch will change the class structure to the following:

QQAbstractDialog <- QQAbstractFileDialog <- QQFileDialog <- QQPlatformFileDialog

and then register QQPlatformFileDialog as the abstract dialog that
DefaultFileDialog.qml intantiates. This means that a file dialog will have
QML contents ready for backup if the platform file dialog fails.

Change-Id: I65d0fb9c49c501501e441873d6b858475fef5522
Reviewed-by: Shawn Rutledge <shawn.rutledge@digia.com>