Dialogs: let folder be assignable in DefaultFileDialog.qml
The current code creates an explicit binding of property
'folder' to its folder list model. The result is that if
the app sets folder to something else, it would be overwritten
by the binding already upon creation.

This patch will keep the binding, but assign it to the property
directly. That way the app can override its initial value.

Note that since the app cannot manipulate the model directly, and
since functions like dirUp(), dirDown(), click in tableview etc, updates
'folder' directly, the binding seems to be unnecessary in the
first place.

Change-Id: I65e900df9f9305e8bbec4ac401062243d83740aa
Reviewed-by: Shawn Rutledge <shawn.rutledge@digia.com>