Dialogs: redo calculations depending on __maximumDimension before show
Once again the old problem recurs that on Android the primaryScreen
has invalid geometry at the time the dialog's QML implementation is
loaded.  So if the loading occurs before the main window is shown,
the dialog was resized to 0x0 and stayed that way.
The fix is to make __maximumDimension a notifying property and
artificially emit a change signal each time before the dialog is shown.

Task-number: QTBUG-45092
Change-Id: Ie86eb7f7bd3cbe8401ccc98c9cc8829c788a71dd
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>