iOS: Make ApplicationWindow fill entire screen, including under statusbar
We take advantage of the new Qt::WindowFlag that asks the OS to maximize
the window using as much as possible of the available screen geometry,
and then make sure that the contentItem is shifted accordingly so that
it doesn't end up under the statusbar. This allow setting a custom
background image/color on the application window, or adding toolbars
or menu bars that underlay the iOS system status bar.

Change-Id: I6efab2aced7efd274a65f2524ae8c270d20de187
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@theqtcompany.com>