Fix TreeView missing visualization of expander image on insert or remove
The current implementation doesn't emit the datachanged signal on the
hasChildren role during insertion or remove of rows. This could cause the
missing visualization of the expander images in the view. This happen
when a new row is inserted inside a leaf node or when a previously not leaf
node becomes a leaf because its last child has been removed.

Change-Id: I4145cb6bd58fe2efeb38c5c14f10657d956fc6ed
Reviewed-by: Filippo Cucchetto <filippocucchetto@gmail.com>
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@theqtcompany.com>