Dismiss a popup window when touch occurs outside
QQuickWindow synthesizes mouse events only for touch points
which occur inside the window. This did not work correctly for
popup windows, which expect a mouse press outside the window
to dismiss the popup.

The workaround is to specifically wait for a TouchBegin outside
the popup to dismiss it.

Task-number: QTBUG-45079
Change-Id: I232220a3fe48c3193299a6a8313a6b9010dd4a53
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>
Reviewed-by: Shawn Rutledge <shawn.rutledge@digia.com>