Dialogs: add shortcuts even when they point to non-existing dirs
Currently, shortcuts will only be added if the directories they point
to exists. Since QStandardPaths::writableLocation() is documented to
return paths that might not yet exist, bindings like the following
will cause warnings on e.g iOS:

FileDialog { folder: shortcuts.pictures; selectExisting: false }

This patch will change this, so that we always add all shortcuts, to
not cause trouble for bindings. That way the app can also take
the necessary steps to check and create a shortcut up front before
showing the dialog.

Change-Id: Iba85710c87ae95e82728695cf958e1bf4394bc7c
Reviewed-by: Shawn Rutledge <shawn.rutledge@digia.com>