Menu: Separate dismiss and destroy actions
This avoids issues when the action is triggered from a Menu and the slot
ends up processing pending events, effectively deleting an object while
one of its QML signal handlers is being executed. The reason being that
we used to call deleteLater() on the menu popup window while still in
the mouse event handler.

Now, we do the same thing in three separate steps. 1. Close/dismiss the
menu popups, 2. trigger the action, and 3. delete the popups. This keeps
the menu popups and their contents alive until we return from the action
triggered handler. We also need to take care of manually destroying any
popup we may create. Finally, the menu content creation in Menu.qml had
to be tweaked since we shouldn't rely on the popup visibility anymore.

Task-number: QTBUG-45182
Change-Id: I9f1155bbf74dd3353c6c4066a24abf1cd2c3a283
Reviewed-by: Caroline Chao <caroline.chao@theqtcompany.com>
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>