MenuBar top level menus are not closed on click after being opened
This fix revert partially what was done for resolving QTBUG-40391.
Infact the change QTBUG-40391 caused the propagation of the events
that caused the dismiss of a popupMenu. This in turn caused this bug
where the click that should close menu is forwarded instead of being
swallowed. However for supporting the behavior of context menu
outside menubar we added some cases where the developer expect
to receive the event that dismissed the popup. The use case is to
reopen a dismissed popup on right click (see windows right click
behavior on right clicks)

Task-number: QTBUG-45315
Change-Id: If5a181b65ed9d879521ac20de577510908125169
Reviewed-by: Filippo Cucchetto <filippocucchetto@gmail.com>
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@theqtcompany.com>