TextEdit: don't clear selection from EditMenu
The code as it stood ensured that only one input
field had a text selection at a time. This is not
correct, several text edit fields are allowed to have
a selection simultaneously.
The code that cleared the selection was found in the
updateSelectionItem code.

iOS is the only platform using the selectionItem.
But looking closer at how selections work on iOS, we
don't really need to track the selectionItem after all.
It's fine that you need to gain focus on a line edit /
text edit before you open the edit menu. Moreover, we
also lacked the possibility to open the selection menu
when just clicking on a focused line edit. So this
patch will therefore just remove all the code that
had todo with the selectionItem, and adjust a bit how
selections work on iOS.

Task-number: QTBUG-41375
Change-Id: I0f3951cb7f53e5adae8d8753a7058ee316825921
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>