Don't let popups cross screen boundaries.
This was working on Qt 5.4.1 but got broken by de9145dcd90.
Can't see nothing fundamentally wrong with that commit, problem
was that the code to reposition the popup assumed the contentItem's hight or
width would change every time and connected to hightChanged/widthChanged to
trigger the repositioning. Since de9145dcd90 those signals are only emitted the
first time the popup is shown.

Call updateSize() explicitly.

Task-number: QTBUG-47092
Change-Id: I303bfae80e5cb72e0c24e26365b8d0faa9ddf400
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@theqtcompany.com>