Allow for styles to provide own EditMenu
Allow for styles to define their own EditMenu component, to provide
cut/copy/paste actions with the platform style.
In order to be able to provide mobile-optimized controls on Linux
systems, use the boolean env var QT_MOBILE_PLATFORM in this case.
This allows on a plain Linux system to offer controls with the
touch friendly text selection handles and the little inline
cut/copy/paste menu

Change-Id: Id9013b4386423938735390325a003fcccdcfd598
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@theqtcompany.com>