Fix TypeErrors on control destruction with singleton style components
When destroying a control whose style points to a component defined in
a QML singleton, there are errors about the parent of the panel being
null.

This patch moves the style loader out of the panel loader, which fixes
the errors. However, as mentioned in the comments of the bug report,
the destruction order seems to be unchanged, so it is unclear why this
works.

Change-Id: I4a9e33f2ddf6cfa0f069991367b87f91bec99410
Task-number: QTBUG-38422
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>