Menus: Clean popup hide and destroy logic
When a menu popup gets closed it usually needs to be destroyed right away
since we don't recycle its contents. There is an exception, however, and
it's when he user triggers a menu item.

In this case, we need to proceed in three steps. First, we hide the menu
popup, then we emit the triggered signal, and when that one returns, the
menu contents can be disposed. If we did all in a single step, we may end
up with a crash since we don't support deleting a QtQuick item while it's
running a signal handler. Delayed deletions don't work either in the case
when the triggered handler ends up running the event loop.

Task-number: QTBUG-45182
Task-number: QTBUG-47682
Task-number: QTBUG-48382
Change-Id: Ic39717e09f38df602f641250cd81cf4931863db6
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>