BasicTableView: Use ListView's overlay header for the table header
Previoulsy we manually positioned the table header over the table
contents, which required some manual tweakings. Since 5.4, ListView
support overlay headers that we can use them to clean our code in
BasicTableView.

There is, however, a small issue on Mac where the vertical scrollbar
(specially in non-transient mode) is displayed below the table header.
Since the header can't span beyond the ListView boundaries, we need
to account for the vertical scrollbar width in the ListView's content
width.

This patch also updates the Desktop style to set the transientScrollBars
property and some small tweaks to the Base style.

Change-Id: I4433d0703328f23eadb0ae4563fbef5bba0168db
Task-number: QTBUG-34344
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>