Dialog: allow proper resize and default size for content
Currently Dialog use implicitly positioned items as content. This makes
impossible to use item, that can be properly resized with change of
dialog window size and has proper default size (set implicit width and
height).

This change made it use implicit size if there is one item or fallback
to previous algorithm, that use childrenRect for explicitly sized and
positioned items, so it won't break behavior of the existing
applications.

[ChangeLog][Dialog] Proper resize and default size for content

Task-number: QTBUG-49058
Change-Id: I7fa4da7d5fd39d47f60ba9f43c88f78e75739b0a
Initial-patch-by: Nikita Krupenko <krnekit@gmail.com>
Reviewed-by: Nikita Krupenko <krnekit@gmail.com>
Reviewed-by: Shawn Rutledge <shawn.rutledge@theqtcompany.com>