Menubar popups should handle only release events of a previous press event
When a popup is brought up by a press event for a menu
inside the menubar it should not handle the relative
next release event if it falls outside the popup.
Before forwarding the release event we first check if a
previous press event was seen, if not we simply discard
the event.

Task-number: QTBUG-47295
Task-number: QTBUG-45117
Change-Id: I632fab0a3abfdfc9872f85f99f9d7f50d41526cc
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>