iOS: let platform plugin populate edit menus
When showing an edit menu on iOS, UIKit will always populate
the menu with some default system menu items.
This means that if controls e.g adds a menu item "cut" to the
menu, it might appear twize since UIKit might add one as
well. To avoid this, we filter away menu items added by Qt if
we see that the shortcut assigned corresponds to one of the
UIKit items. As such, we rely on the menu items in the edit
menu to have shortcuts assigned.

But this strategy turns out to have a flaw. A specific
shortcut can only map to one control inside a window. So if
a control creates a shortcut "copy", no-one else inside
that window can do the same. So if you have several line
edits, all with their own edit menu, you will end up with
many menu items using the same shortcuts. And this will
cause warnings and problems when Qt later tries to find a
shortcut to trigger.

Since it turns out the using shortcuts internally for our own
controls is a bad idea in the first place (it steals away the
possibility for the application to use them for something else), and
since we're anyway moving into the direction of doing text editing
directly from the platform plugin, this change will start by
giving full control over the contents of the edit menu to the ios
plugin. This will remove usage of shortcuts, and therefore fix
the related problems in the same go.

Change-Id: Ib8a6f54d444e756fee7328e06a4b83e778cb1022
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>
Reviewed-by: Markus Goetz (Woboq GmbH) <markus@woboq.com>